Refactor authentication handling by moving related logic from movie slice and API to auth slice and API. Update imports in components and pages accordingly. Remove unused movie card component and HTTP utility. Clean up movie API by removing obsolete user-related endpoints and types. Enhance environment configuration with additional documentation contact fields.
