--- client/src/lib/http.ts
@@ -1,8 +1,0 @@
-import axios from "axios";
-
-export const serverApi = axios.create({
-  baseURL: "http://localhost:3000/api",
-  headers: {
-    "Content-Type": "application/json",
-  },
-});
